Merge branch 'dev/wns-62/script_to_delete_users' into 'develop'

See merge request freelance-website/back!21

+++ app/base/models.py
@@ -3,29 +3,28 @@ from django.utils import timezone
 
 from .managers import BaseModelManager
 
+
 class BaseModel(models.Model):
-  created_at = models.DateTimeField(auto_now_add=True, blank=True, null=True)
-  updated_at = models.DateTimeField(auto_now=True, blank=True, null=True)
-  deleted_at = models.DateTimeField(blank=True, null=True)
-  is_deleted = models.BooleanField(default=False)
+    created_at = models.DateTimeField(auto_now_add=True, blank=True, null=True)
+    updated_at = models.DateTimeField(auto_now=True, blank=True, null=True)
+    deleted_at = models.DateTimeField(blank=True, null=True)
+    is_deleted = models.BooleanField(default=False)
 
-  objects = BaseModelManager()
-  all_objects = BaseModelManager(with_deleted=True)
+    objects = BaseModelManager()
+    all_objects = BaseModelManager(with_deleted=True)
 
-  class Meta:
-    abstract = True
-  
+    class Meta:
+        abstract = True
 
-  def delete(self):
-    self.deleted_at = timezone.now()
-    self.is_deleted = True
-    self.save()
+    def delete(self, soft=True):
+        if soft:
+            self.deleted_at = timezone.now()
+            self.is_deleted = True
+            self.save()
+        else:
+            super(BaseModel, self).delete()
 
-  
-  def hard_delete(self):
-    super(BaseModel, self).delete()
-
-  def restore(self):
-    self.deleted_at = None
-    self.is_deleted = False
-    self.save()
+    def restore(self):
+        self.deleted_at = None
+        self.is_deleted = False
+        self.save()
